backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "google-genai>=1.27.0",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-unstructured>=0.1.6",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "stripe>=12.3.0",
    "ulid>=1.1",
    "uvicorn>=0.34.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx"
]
